lxd/apparmor: Move dnsmasq functions
