+++ lxd/apparmor/network.go
@@ -1,50 +1,16 @@
 package apparmor
 
 import (
-	"crypto/sha256"
-	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/lxc/lxd/lxd/state"
-	"github.com/lxc/lxd/shared"
 )
 
 // Internal copy of the network interface.
 type network interface {
 	Name() string
-}
-
-// DnsmasqProfileName returns the AppArmor profile name.
-func DnsmasqProfileName(n network) string {
-	path := shared.VarPath("")
-	name := fmt.Sprintf("%s_<%s>", n.Name(), path)
-
-	// Max length in AppArmor is 253 chars.
-	if len(name)+12 >= 253 {
-		hash := sha256.New()
-		io.WriteString(hash, name)
-		name = fmt.Sprintf("%x", hash.Sum(nil))
-	}
-
-	return fmt.Sprintf("lxd_dnsmasq-%s", name)
-}
-
-// dnsmasqProfileFilename returns the name of the on-disk profile name.
-func dnsmasqProfileFilename(n network) string {
-	name := n.Name()
-
-	// Max length in AppArmor is 253 chars.
-	if len(name)+12 >= 253 {
-		hash := sha256.New()
-		io.WriteString(hash, name)
-		name = fmt.Sprintf("%x", hash.Sum(nil))
-	}
-
-	return fmt.Sprintf("lxd_dnsmasq-%s", name)
 }
 
 // NetworkLoad ensures that the network's profiles are loaded into the kernel.
@@ -100,27 +66,4 @@ func NetworkUnload(state *state.State, n network) error {
 // NetworkDelete removes the profiles from cache/disk.
 func NetworkDelete(state *state.State, n network) error {
 	return deleteProfile(state, dnsmasqProfileFilename(n))
-}
-
-// dnsmasqProfile generates the AppArmor profile template from the given network.
-func dnsmasqProfile(state *state.State, n network) (string, error) {
-	rootPath := ""
-	if shared.InSnap() {
-		rootPath = "/var/lib/snapd/hostfs"
-	}
-
-	// Render the profile.
-	var sb *strings.Builder = &strings.Builder{}
-	err := dnsmasqProfileTpl.Execute(sb, map[string]interface{}{
-		"name":        DnsmasqProfileName(n),
-		"networkName": n.Name(),
-		"varPath":     shared.VarPath(""),
-		"rootPath":    rootPath,
-		"snap":        shared.InSnap(),
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return sb.String(), nil
 }
