+++ lxd/apparmor/network_dnsmasq.go
@@ -1,7 +1,14 @@
 package apparmor
 
 import (
+	"crypto/sha256"
+	"fmt"
+	"io"
+	"strings"
 	"text/template"
+
+	"github.com/lxc/lxd/lxd/state"
+	"github.com/lxc/lxd/shared"
 )
 
 var dnsmasqProfileTpl = template.Must(template.New("dnsmasqProfile").Parse(`#include <tunables/global>
@@ -59,3 +66,55 @@ profile "{{ .name }}" flags=(attach_disconnected,mediate_deleted) {
 {{- end }}
 }
 `))
+
+// dnsmasqProfile generates the AppArmor profile template from the given network.
+func dnsmasqProfile(state *state.State, n network) (string, error) {
+	rootPath := ""
+	if shared.InSnap() {
+		rootPath = "/var/lib/snapd/hostfs"
+	}
+
+	// Render the profile.
+	var sb *strings.Builder = &strings.Builder{}
+	err := dnsmasqProfileTpl.Execute(sb, map[string]interface{}{
+		"name":        DnsmasqProfileName(n),
+		"networkName": n.Name(),
+		"varPath":     shared.VarPath(""),
+		"rootPath":    rootPath,
+		"snap":        shared.InSnap(),
+	})
+	if err != nil {
+		return "", err
+	}
+
+	return sb.String(), nil
+}
+
+// DnsmasqProfileName returns the AppArmor profile name.
+func DnsmasqProfileName(n network) string {
+	path := shared.VarPath("")
+	name := fmt.Sprintf("%s_<%s>", n.Name(), path)
+
+	// Max length in AppArmor is 253 chars.
+	if len(name)+12 >= 253 {
+		hash := sha256.New()
+		io.WriteString(hash, name)
+		name = fmt.Sprintf("%x", hash.Sum(nil))
+	}
+
+	return fmt.Sprintf("lxd_dnsmasq-%s", name)
+}
+
+// dnsmasqProfileFilename returns the name of the on-disk profile name.
+func dnsmasqProfileFilename(n network) string {
+	name := n.Name()
+
+	// Max length in AppArmor is 253 chars.
+	if len(name)+12 >= 253 {
+		hash := sha256.New()
+		io.WriteString(hash, name)
+		name = fmt.Sprintf("%x", hash.Sum(nil))
+	}
+
+	return fmt.Sprintf("lxd_dnsmasq-%s", name)
+}
